requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
motor
scikit-learn
python-multipart
nltk
matplotlib
pandas
joblib
python-docx
